Add a list command to list all files under a path, using the higher performance recursive list operation.
Prints out some statistics too.
If you test this against s3a//landsat-pds/ you can see the paginated behaviour of the list operation;
the log is always stuck at a -9999 value at the bottom of the page because that's when the client pauses to ask for another thousand
entries.
We should think about doing that page fetch in advance, before you get to the final entry.
At the cost of more local storage, it will be faster.

Something like: first next() call: fetch first page, return the first value, then do an async fetch of the second page.
As every end of page is reached, the next background page fetch is waited on, a new fetch queued, and the result of the
latest one handed back.