StoreDiag: downgrade required JARs; probe for implementing FS.
Allows for probes for filesystems with shaded dependencies.

hadoop jar $CLOUDSTORE storediag -o -r abfs://bucket@account/

The implementing class is probed for using the FileSystem method
FileSystem.getFileSystemClass(schema); this ensures that
the codepath for self-registering filesystems is followed,
and the final implementation class discovered this way
is determined.

It also tries to print a useful message on a failure.

+ removed the storediagjob/wc classes, which weren't actively
used/maintained.