HADOOP-19256. S3A: Support Conditional Writes
Amazon S3 now supports conditional overwrites, which can be be used
when creating files through the createFile() API with two new
builder options:

          fs.option.create.conditional.overwrite

Write if and only if there is no object at the target path.
This is an atomic PUT-no-overwrite, checked in close(), not create().

          fs.option.create.conditional.overwrite.etag

Write a file if and only if it is overwriting a file with a specific
etag.

If the "fs.s3a.performance.flags" enumeration includes the flag "create"
then file creation will use conditional creation to detect and reject
overwrites.

The configuration option "fs.s3a.create.conditional.enabled"
can be set to false to disable these features on third-party stores.

Contributed by Diljot Grewal, Saikat Roy and Steve Loughran